B2B/dashboard/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.0",
    "bcrypt>=4.2.0",
    "cryptography>=47.0.0",
    "dnspython>=2.8.0",
    "fastapi>=0.136.1",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyjwt[crypto]>=2.8.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.46.0",
    "xlsxwriter>=3.2.0",
]
